# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Grid Origination Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: market
          in: query
          schema:
            type: string
            enum: [ERCOT, CAISO, PJM]
        - name: assetType
          in: query
          schema:
            type: string
            enum: [solar, wind, storage, solar_storage, wind_storage]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, under_review, contracted]
        - name: minScore
          in: query
          schema:
            type: number
        - name: maxScore
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 5000
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of candidates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a new candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a single candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "200":
          description: Updated candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /queue-projects:
    get:
      operationId: listQueueProjects
      tags: [queue-projects]
      summary: List interconnection queue projects
      parameters:
        - name: market
          in: query
          schema:
            type: string
            enum: [ERCOT, CAISO, PJM]
        - name: status
          in: query
          schema:
            type: string
        - name: fuelType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 2000
      responses:
        "200":
          description: Queue projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueProject"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    Candidate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        market:
          type: string
          enum: [ERCOT, CAISO, PJM]
        assetType:
          type: string
          enum: [solar, wind, storage, solar_storage, wind_storage, hydro, biomass, geothermal, natural_gas, nuclear]
        status:
          type: string
          enum: [active, inactive, under_review, contracted]
        capacityMw:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        county:
          type: string
        state:
          type: string
        interconnectionNode:
          type: string
        pricingHubNode:
          type: string
        estimatedLcoe:
          type: number
        offtakePriceMwh:
          type: number
        overallScore:
          type: number
        priceScore:
          type: number
        locationScore:
          type: number
        curtailmentScore:
          type: number
        interconnectionScore:
          type: number
        regulatoryScore:
          type: number
        financialScore:
          type: number
        environmentalScore:
          type: number
        gridStabilityScore:
          type: number
        demandProximityScore:
          type: number
        developmentRiskScore:
          type: number
        notes:
          type: string
        recEligible:
          type: boolean
          description: Whether this asset type earns RECs (solar, wind, hydro, geothermal, biomass = eligible; storage, gas, nuclear = not)
        annualRecMwh:
          type: number
          description: Estimated annual REC generation in MWh (capacity × capacity_factor × 8760)
        recPricePerMwh:
          type: number
          description: Market benchmark REC price in $/MWh (ERCOT TRC ~$1.50, CAISO WREGIS ~$7-12, PJM $2-120)
        annualRecValueUsd:
          type: number
          description: Estimated annual REC revenue in USD
        lifetimeRecValue20yr:
          type: number
          description: 20-year undiscounted REC value in USD
        recMarketLabel:
          type: string
          description: Human-readable label for the REC market (e.g. Texas TRC, CA WREGIS RPS)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - market
        - assetType
        - status
        - capacityMw
        - latitude
        - longitude
        - overallScore
        - createdAt
        - updatedAt
    CandidateInput:
      type: object
      properties:
        name:
          type: string
        market:
          type: string
          enum: [ERCOT, CAISO, PJM]
        assetType:
          type: string
          enum: [solar, wind, storage, solar_storage, wind_storage]
        status:
          type: string
          enum: [active, inactive, under_review, contracted]
        capacityMw:
          type: number
        latitude:
          type: number
        longitude:
          type: number
        county:
          type: string
        state:
          type: string
        interconnectionNode:
          type: string
        pricingHubNode:
          type: string
        estimatedLcoe:
          type: number
        offtakePriceMwh:
          type: number
        priceScore:
          type: number
        locationScore:
          type: number
        curtailmentScore:
          type: number
        interconnectionScore:
          type: number
        regulatoryScore:
          type: number
        financialScore:
          type: number
        environmentalScore:
          type: number
        gridStabilityScore:
          type: number
        demandProximityScore:
          type: number
        developmentRiskScore:
          type: number
        notes:
          type: string
      required:
        - name
        - market
        - assetType
        - capacityMw
        - latitude
        - longitude
...
    QueueProject:
      type: object
      properties:
        id:
          type: integer
        projectName:
          type: string
        market:
          type: string
          enum: [ERCOT, CAISO, PJM]
        queueId:
          type: string
        fuelType:
          type: string
        capacityMw:
          type: number
        status:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        county:
          type: string
        state:
          type: string
        interconnectionNode:
          type: string
        requestDate:
          type: string
          format: date-time
        studyGroupPhase:
          type: string
        withdrawalDate:
          type: string
          format: date-time
        recEligible:
          type: boolean
          description: Whether this fuel type earns RECs
        annualRecMwh:
          type: number
          description: Estimated annual REC generation in MWh
        recPricePerMwh:
          type: number
          description: Market benchmark REC price in $/MWh
        annualRecValueUsd:
          type: number
          description: Estimated annual REC revenue in USD
        lifetimeRecValue20yr:
          type: number
          description: 20-year undiscounted REC value in USD
        recMarketLabel:
          type: string
          description: Human-readable label for the REC market
      required: [id, projectName, market, fuelType, capacityMw, status]
...