# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grid Origination Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CandidateAssetType } from "./candidateAssetType";
import type { CandidateMarket } from "./candidateMarket";
import type { CandidateStatus } from "./candidateStatus";

export interface Candidate {
  id: number;
  name: string;
  market: CandidateMarket;
  assetType: CandidateAssetType;
  status: CandidateStatus;
  capacityMw: number;
  latitude: number;
  longitude: number;
  county?: string;
  state?: string;
  interconnectionNode?: string;
  pricingHubNode?: string;
  estimatedLcoe?: number;
  offtakePriceMwh?: number;
  overallScore: number;
  priceScore?: number;
  locationScore?: number;
  curtailmentScore?: number;
  interconnectionScore?: number;
  regulatoryScore?: number;
  financialScore?: number;
  environmentalScore?: number;
  gridStabilityScore?: number;
  demandProximityScore?: number;
  developmentRiskScore?: number;
  notes?: string;
  /** Whether this asset type earns RECs (solar, wind, hydro, geothermal, biomass = eligible; storage, gas, nuclear = not) */
  recEligible?: boolean;
  /** Estimated annual REC generation in MWh (capacity × capacity_factor × 8760) */
  annualRecMwh?: number;
  /** Market benchmark REC price in $/MWh (ERCOT TRC ~$1.50, CAISO WREGIS ~$7-12, PJM $2-120) */
  recPricePerMwh?: number;
  /** Estimated annual REC revenue in USD */
  annualRecValueUsd?: number;
  /** 20-year undiscounted REC value in USD */
  lifetimeRecValue20yr?: number;
  /** Human-readable label for the REC market (e.g. Texas TRC, CA WREGIS RPS) */
  recMarketLabel?: string;
  createdAt: Date;
  updatedAt: Date;
}