# lib/api-zod/src/generated/types/queueProject.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grid Origination Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { QueueProjectMarket } from "./queueProjectMarket";

export interface QueueProject {
  id: number;
  projectName: string;
  market: QueueProjectMarket;
  queueId?: string;
  fuelType: string;
  capacityMw: number;
  status: string;
  latitude?: number;
  longitude?: number;
  county?: string;
  state?: string;
  interconnectionNode?: string;
  requestDate?: Date;
  studyGroupPhase?: string;
  withdrawalDate?: Date;
  /** Whether this fuel type earns RECs */
  recEligible?: boolean;
  /** Estimated annual REC generation in MWh */
  annualRecMwh?: number;
  /** Market benchmark REC price in $/MWh */
  recPricePerMwh?: number;
  /** Estimated annual REC revenue in USD */
  annualRecValueUsd?: number;
  /** 20-year undiscounted REC value in USD */
  lifetimeRecValue20yr?: number;
  /** Human-readable label for the REC market */
  recMarketLabel?: string;
}